### 2024-01-02 #6176 @editorialbot
[PRE REVIEW]: growR: R Implementation of the Vegetation Model ModVege
**Submitting author:** @kuadrat (<a href="http://orcid.org/0000-0001-5523-6924">Kevin Kramer</a>)
**Repository:** https://github.com/kuadrat/growR
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.2.0
