### 2024-01-02 #6176 @kthyng
Hi @kuadrat and thanks for your submission! I'll add it to our waitlist, but in the meantime please check:
* you have two license files in your repo, please keep only one and and OSI-approved license.
* Add a statement of need section to your paper
* suggest 5 reviewers from our database or your extended but not conflicting network without actually pinging them with "@"
