### 2024-01-03 #6176 @kuadrat
Hi @kthyng, happy New Year and thanks for taking up my submission!

Regarding the points you mention:

- The existence of two license files is a consequence of different expectations by github and CRAN. The issue is discussed, [for example on stackoverflow](https://stackoverflow.com/questions/43550479/how-to-satisfy-both-cran-and-github-license-file-naming-requirements). This seems to be a common problem for submissions to JOSS, see e.g. https://github.com/jessecambon/tidygeocoder/issues/128 and https://github.com/kadyb/rgugik/issues/61. In those cases the conclusion was that nothing could be done about it...
- The first section is called *Summary and Statement of Need*, would that check the box for containing a statement of need?. Otherwise, I am happy to rename it to just *Statement of Need*.
- Potential reviewers just based on their areas of expertise as stated in the JOSS reviewer database: VEZY, albhasan, jmp75,  shubhamjain15, kanishkan91
