### 2024-01-15 2a83f9d
allow chained model execution, bugfix
