### 2024-01-15 9807eec
doc for `growR_run_loop(independent)`
