### 2024-01-23 cf292b0
JOSS badge in README
