### 2024-02-01 002c49f
Choose SGS method through parameters
