+++ R/modvegesite.R
@@ -12,7 +12,6 @@ initial_run_variables = list(
   BMGRp = NULL,
   BMDRp = NULL,
   BMDVp = NULL,
-  STp = NULL,
   cBMp = NULL,
   WRp = NULL,
   GROGV = NULL,
@@ -127,7 +126,7 @@ ModvegeSite = R6Class(
 #' @field cut_during_growth_preriod Boolean to indicate whether a cut 
 #'   occurred during the growth period, in which case reproductive growth is 
 #'   stopped.
-     cut_during_growth_preriod = NULL,
+      cut_during_growth_preriod = NULL,
 #' @field last_DOY_for_initial_cut [autocut] Start cutting after this DOY, 
 #'   even if yield target is not reached.
       last_DOY_for_initial_cut = 150,
@@ -172,7 +171,7 @@ ModvegeSite = R6Class(
 #' @field stubble_height float. Minimum height the grass can assume. The 
 #'    biomass will not fall below that height. This can and should therefore 
 #'    be smaller than `self$cut_height`.
-      stubble_height = 0.05,
+      stubble_height = 0.02,
 
   #-Public-methods----------------------------------------------------------------
 
@@ -188,6 +187,7 @@ ModvegeSite = R6Class(
         self$site_name = site_name
         self$run_name = run_name
         self$parameters = parameters
+        self$set_SGS_method(self$parameters$SGS_method)
 
         self$state_variable_names = names(initial_state_variables)
         self$n_state_variables = length(self$state_variable_names)
@@ -220,6 +220,30 @@ ModvegeSite = R6Class(
           stop("ModvegeSite$management has not been set.")
         } else {
           return(self$management)
+        }
+      },
+
+      #' @description Choose which method to be used for determination of SGS
+      #' 
+      #' Options for the determination of the start of growing season (SGS) are:
+      #' \describe{
+      #'   \item{MTD}{Multicriterial thermal definition, 
+      #'     [start_of_growing_season_mtd()]}
+      #'   \item{simple}{Commonly used, simple SGS definition based on 
+      #'     temperature sum, [start_of_growing_season()]}
+      #' }
+      #'
+      #' @param method str Name of the method to use. Options: "MTD", "simple".
+      #' @return none
+      #'
+      #' @seealso [start_of_growing_season_mtd()], [start_of_growing_season()]
+      #'
+      set_SGS_method = function(method) {
+        if (method %in% private$SGS_options) {
+          private$SGS_method = method
+        } else {
+          warning(sprintf("Unrecognized SGS method `%s`. Use one of %s.",
+                  paste(private$SGS_options, collapse = ", ")))
         }
       },
 
@@ -363,6 +387,7 @@ ModvegeSite = R6Class(
         self$management = management
         self$year = year
         private$initialize_state_variables()
+        private$calculate_temperature_sum()
         private$get_start_of_growing_season()
 
         # Determine whether cuts are to be read from input or to be 
@@ -638,6 +663,8 @@ ModvegeSite = R6Class(
     vars_to_exclude = c("OMDDV", "OMDDR"),
     current_DOY = 1,
     REP_ON = NULL,
+    SGS_method = "MTD",
+    SGS_options = c("MTD", "simple"),
     # List of labels to use for different variables when plotting.
     ylabels  =  list(AgeGV = "AgeGV (degree days)",
                    AgeGR = "AgeGR (degree days)",
@@ -723,9 +750,30 @@ ModvegeSite = R6Class(
       self[["BMGRp"]]  = P$BMGR0
       self[["BMDRp"]]  = P$BMDR0
       self[["BMDVp"]]  = P$BMDV0
-      self[["STp"]]    = P$ST0
       self[["cBMp"]]   = P$cBM0
       self[["WRp"]]    = P$WR0
+    },
+
+    ## @description
[…]
