+++ R/parameters.R
@@ -42,7 +42,8 @@ parameter_defaults = list(
   OMDDR = 0.4,
   CO2_growth_factor = 0.5,
   crop_coefficient = 1.15,
-  senescence_cap = 0.7
+  senescence_cap = 0.7,
+  SGS_method = "MTD"
 )
 # Create a list that can be inserted into R6 class to create many fields
 initial_condition_names = names(initial_conditions)
