### 2024-02-12 #6260 @kuadrat
Regarding license (as I can see this is an unchecked box in the review checklist of @shubhamjain15 ), please have a look at the discussion that has already occurred during pre-review and links therein:

https://github.com/openjournals/joss-reviews/issues/6176#issuecomment-1874964381

In short: At the moment it does not seem to be possible to simultaneously satisfy CRAN's and github's conventions regarding naming and structure of license files. Since the main avenue of distribution is via CRAN, the choice has been made to follow their conventions. The conclusion in JOSS reviews for other R packages has been that this complies with JOSS (see linked comment).
