### 2024-02-12 ecbd52c
stubble height to parameters
