+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: growR
 Type: Package
-Version: 1.2.0.9003
+Version: 1.2.0.9004
 Date: 2023-11-21
 Authors@R: person(
     given = "Kevin",
