### 2024-02-12 #6260 @RobLBaker
Really cool work!

I missed the **statement of need** in the Documentation. I am assuming this is separate from the **statement of need** in the Software paper. Perhaps I'm overlooking it somewhere - if you could please point me in the right direction or confirm that these two statements of need are one and the same, I'd be happy to check that last box.
