+++ README.md
@@ -22,12 +22,56 @@ tools for analysing model outputs and performance.
 
 ## Contents
 
-1. [Installation](#installation)
-2. [Getting Started](#getting-started)
+1. [What is this for?](#what-is-this-for?)
+2. [Installation](#installation)
+3. [Getting Started](#getting-started)
 4. [Contributing](#contributing)
 5. [Contact](#contact)
-3. [Glossary](#glossary)
+6. [Glossary](#glossary)
 7. [Footnotes and References](#footnotes-and-references)
+
+## What is this for?
+
+This `R` package allows simulation of grass growth.
+
+### Why simulate grass growth?
+
+Grasslands constitute one of Earth's most widespread terrestrial 
+ecosystems[^3] and a core element in global agriculture, providing roughly 
+half the feed inputs for global livestock systems 
+[^4].
+Beside their contribution to global food production, they provide a catalogue 
+of other ecosystem services, such as water flow and erosion regulation, 
+pollination service, carbon sequestration and climate regulation 
+[^3].
+The latter have become particularly important in light of anthropogenic 
+climate change [^5].
+
+Understanding the functioning of grassland ecosystems and their responses to 
+external changes is therefore of significant interest.
+Vegetation models provide a powerful platform for such studies.
+
+### How does this compare to other grass and vegetation models?
+
+The number of grassland models is large and ever-growing.
+We can therefore not give a comprehensive list, but will try to make a couple 
+of representative comparisons to illustrate where `growR` has its niche.
+For the most part, an advantage of `growR` over other, similar models and 
+their implementations is its distribution as `R` package via CRAN.
+
+- The [Hurley Pasture Model](https://sites.massey.ac.nz/hurleypasturemodel/hurley-pasture-model/) [^6]
+  is a detailed mechanistic model for managed pastures. It is implemented in 
+  the *Advanced continuous simulation language (ACSL)* and the source code is 
+  available on request.
+- [BASGRA](https://github.com/davcam/BASGRA/) [^7] 
+  and its descendant [BASGRA_N](https://github.com/MarcelVanOijen/BASGRA_N) 
+  [^8] are multi-year grassland models which include tiller dynamics.
+  They are also implemented in `R` with the source code freely available. 
+  However, they do not come packaged, as `growR` does.
+- PROGRASS [^9] was developed to capture the interactions in grass/clover 
+  mixtures. As of this writing, no accessible implementation was found.
+- The focus of PaSim [^10] is the investigation of livestock production, 
+  which is not directly covered in `growR`, under climate change conditions.
 
 ## Installation
 
@@ -190,6 +234,45 @@ und Eric Mosimann. „Testing the Ability of a Simple Grassland Model to
 Simulate the Seasonal Effects of Drought on Herbage Growth“. Field Crops 
 Research 187 (Februar 2016): 12–23. 
 [doi:10.1016/j.fcr.2015.12.008](https://doi.org/10.1016/j.fcr.2015.12.008).
+
+[^3]: Zhao, Yuanyuan, Zhifeng Liu, and Jianguo Wu. “Grassland Ecosystem 
+Services: A Systematic Review of Research Advances and Future Directions.” 
+Landscape Ecology 35, no. 4 (April 1, 2020): 793–814. 
[…]
