### 2024-02-14 #6260 @kuadrat
Dear @shubhamjain15 , @RobLBaker ,

thank you very much for your review efforts! I'm very glad to hear that the package is received mostly positively.

@RobLBaker I have extended the README with a *What is this for?* section on the [joss branch](https://github.com/kuadrat/growR/tree/joss). Once I merge this into `master`, the [docs](https://kuadrat.github.io/growR/) will be updated accordingly. It is essentially a slightly less formal rewording of relevant paragraphs in the paper. Please let me know what you think.
