### 2024-02-14 #6260 @mikemahoney218
@kuadrat , looks like that invalid DOI should be 10.1017/9781009325844.007 (I think the `.714` is the issue) -- any chance you can also update that?
