+++ joss_paper/paper.bib
@@ -58,7 +58,7 @@
   address = {Cambridge, UK and New York, USA},
   pages = {713-906},
   ISBN = {9781009325844},
-  DOI = {10.1017/9781009325844.007.714},
+  DOI = {10.1017/9781009325844.007},
   year = {2022},
   type = {Book Section}
 }
