+++ pkgdown.yml
@@ -10,7 +10,7 @@ articles:
   weather_parameters: children/weather_parameters.html
   growR: growR.html
   parameter_descriptions: parameter_descriptions.html
-last_built: 2024-01-23T07:20Z
+last_built: 2024-02-14T14:29Z
 urls:
   reference: https://kuadrat.github.io/growR/reference
   article: https://kuadrat.github.io/growR/articles
