### 2024-02-14 #6260 @kuadrat
> Awesome! Thank you so much @shubhamjain15 and @RobLBaker for your thoughtful reviews here. As both reviewers are recommending acceptance, I think we're good to move on to final processing steps!
>
> @kuadrat , at this point could you:
>
>     * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
>
>     * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
>
>     * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
>
>     * [ ]  Please list the DOI of the archived version here.
>
>
> I can then move forward with recommending acceptance of the submission.

- As you've already seen, the broken DOI has been fixed.
- A tagged release has been created: [growR-1.2.0-JOSS](https://github.com/kuadrat/growR/releases/tag/growR-1.2.0-JOSS)
- Source code has been archived at [Zenodo](https://zenodo.org/records/10658194) under DOI: [10.5281/zenodo.10658193](https://zenodo.org/doi/10.5281/zenodo.10658193)
