### 2024-02-14 #6260 @danielskatz
It's saying that an ISBN element in the bib file has a length of 9, and ISBNs should be 10 characters, so this isn't a valid ISBN.
