### 2024-02-14 #6260 @kuadrat
Upon inspection, all of the ISBN's in the .bib file were 13 digit (ISBN-13), except for one (`hugueninElie2017DuengungGrasland`). The latter turns out to be an [ISSN](https://portal.issn.org/).
Do you think it will work if I simply rename the field from ISBN to ISSN or should I rather delete the respective field?
