### 2024-02-14 #6260 @mikemahoney218
Thanks @danielskatz ! After fixing this, will @kuadrat need to generate a new archive and release?

(I don't fully know @kuadrat -- it seems to me that setting the field to `issn` would be fine, given the report doesn't have a DOI or ISBN, but I don't actually know!)
