### 2024-02-14 #6260 @kthyng
See if the example paper helps here: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

@mikemahoney218 No need to generate a new archive or release for changes in the paper.
