### 2024-02-14 #6260 @danielskatz
You can try changing this to an ISSN; I don't remember how JOSS processing will handle it, but you can try regenerating the paper to see. I can't see how it would hurt in any case, though we might find out :)

A new archive and release are not needed after this, because JOSS doesn't require the paper source to be in the archive at all, given that the generated paper (pdf, xml, jats) is what is archived by JOSS, in addition to the software that is archived by the author.
