+++ joss_paper/paper.bib
@@ -199,7 +199,7 @@
   pages = {1--22},
   publisher = {{Agrarforschung Schweiz}},
   urldate = {2023-05-24},
-  isbn = {1663-7852},
+  issn = {1663-7852},
 }
 
 @book{soltani2012ModelingPhysiologyCrop,
