### 2024-02-14 #6260 @kuadrat
Alright, thanks @kthyng and @danielskatz for pitching in.
Seeing that there are already other entries with ISSN's, I'm confident that this should now work :)
Latest commit: [26e64d7f](https://github.com/kuadrat/growR/commit/26e64d7f97247075a6c65ea07e61deb5537b0019)
