### 2024-02-14 36db989
Remove duplicate entry in paper.bib
