+++ joss_paper/paper.bib
@@ -1,22 +1,3 @@
-@article{zhao2020GrasslandEcosystemServices,
-  title = {Grassland Ecosystem Services: A Systematic Review of Research Advances and Future Directions},
-  shorttitle = {Grassland Ecosystem Services},
-  author = {Zhao, Yuanyuan and Liu, Zhifeng and Wu, Jianguo},
-  year = {2020},
-  month = apr,
-  journal = {Landscape Ecology},
-  volume = {35},
-  number = {4},
-  pages = {793--814},
-  publisher = {{Springer Netherlands}},
-  issn = {1572-9761},
-  doi = {10.1007/s10980-020-00980-3},
-  urldate = {2023-04-06},
-  copyright = {2020 Springer Nature B.V.},
-  langid = {english},
-  keywords = {read}
-}
-
 @article{herrero2013BiomassUseProduction,
   title = {Biomass Use, Production, Feed Efficiencies, and Greenhouse Gas Emissions from Global Livestock Systems},
   author = {Herrero, Mario and Havl{\'i}k, Petr and Valin, Hugo and Notenbaert, An and Rufino, Mariana C. and Thornton, Philip K. and Bl{\"u}mmel, Michael and Weiss, Franz and Grace, Delia and Obersteiner, Michael},
