### 2024-02-14 #6260 @RobLBaker
Yes, I was able to download and install the package. It appears to function as claimed: I was able to recreate all of the analyses and results from the Tutorial and Calibration vignettes using the example data.
