### 2024-02-15 #6260 @shubhamjain15
Hi Kristen! Yes I was able to install the R package and use the examples in the reference manual and tutorial to test the functionality successfully.
