### 2024-02-19 #6260 @kuadrat
> Congrats on your new publication @kuadrat! Many thanks to editor @mikemahoney218 and reviewers @shubhamjain15 and @RobLBaker for your time, hard work, and expertise!!

Many thanks also to all of you for your voluntary work!

I'm happy this went so smoothly. Maybe a remark on that note: Part of what likely helped making this submission go through without many issues is the fact that the package has been published on [CRAN](https://cran.r-project.org/web/packages/growR/index.html). This means that it has already undergone a number of automated and human-overseen quality checks before being submitted to JOSS.
