### 2024-02-19 #6260 @kthyng
@kuadrat Thanks for that comment and that makes sense. I'll keep my eye out for that being a mitigating (and helpful!) factor on future reviews.

Also if you're interested in getting involved in JOSS, please consider signing up as a reviewer! https://reviewers.joss.theoj.org/
